Moved to using regular stage3 autobuilds, instead of the hardened autobuilds, seeing as the latter are rather likely to be unavailable for a given autobuild date.
The build process now switches to the hardened profile
itself, and rebuilds *all* the packages after recompiling
the toolchain (taking into account that the hardened toolchain
can have different slotted visible versions).

Currently, there should not be a single package that
was not build from source. Previously, that was only partially
the case, because "emerge --depclean" was invoked without
temporarily disabling the "world" packages list.