Separated two kinds of boot: KMS (no VESA) and non-KMS (with VESA).
Removed framebuffer splash from the initramfs stage. KMS boot, which
is the "modern" kind, would need all DRM drivers compiled-in  for
the splash to be available at that stage.

Reenabled the deprecated WIRELESS_EXT, as it is necessary for
NetworkManager, and for IPW 2200 and HERMES drivers.

Added IRQ balancing daemon for SMP systems.