Added carl9170 firmware, updated Florence
Added customm ebuilds for carl9170 firmware, which is unlikely to be added
to portage in the near future, since there are no official binaries, and
building the firmware is a rather convoluted process.

Updated Florence Virtual Keyboard to v0.5.1. Compiling with --with-xrecord
fails, although the reason that XRECORD is turned off by default is not
valid anymore (the relevant X bugs were fixed).