Unsafe Browser now refuses to start when Tor is running
Tor is queried for its status using GETINFO status/circuit-established
(https://gitweb.torproject.org/torspec.git/blob/HEAD:/control-spec.txt).

Since when a network interface goes down, Tor appears to think that
the circuit works indefinitely, Tor is now restarted when an interface
is brought down (it is not stopped, to prevent race conditions and
issues with multi-interface connections).

The check is performed in main user's wrapper script, and is easy to
circumvent. An alternative would be allowing sudo access to user nofw.