Disabled user-level cgroups manipulation
Considering the cgroup patch that was integrated into the kernel,
and dependence on cgroup by critical services like systemd, it is
probably not wise to allow cgroup fs access to the user. E.g., the
user can disable the release of a cgroup. If/when systemd becomes
the standard, it will depend on exclusive control over cgroups.