Implemented Secure Boot signing for EFI GRUB images
Signing is implemented via a certificates chain:
  CA -> Cert-1 -> BOOT*.EFI
Cert-1 is embedded in a PKCS7 PE/COFF signature of BOOT*.EFI
images, and CA is separately bundled to be enrolled into Secure Boot
signatures DB. The CA can be enrolled as a KEK instead, but that is
unnecessary, as it authorizes private key owner to sign executables
that enroll signatures into firmware at run-time.

A 2nd level Cert-1 is used to sign the EFI images instead of CA,
so that other OSes can enroll Cert-1 into DBX blacklist (for whatever
reason) instead of top-level CA, in whih case further images can be
signed with a new Cert-2.