Differentiated between "lost" and readonly boot media
Boot media can become "lost" when it is not recognized as USB 2.0
during early boot (due to missing modules), and is recognized as
such later, in which case, e.g., /dev/sdb disappears and is replaced
with /dev/sdc, while /mnt/boot is still reported as a mountpoint
with a non-existing /dev/block/major:minor device.

This is useful for bug reporting.